+++ Public/Invoke-Fortification.ps1
@@ -26,7 +26,7 @@ function Invoke-Fortification {
         [string]$VaultName = 'PSGuerrilla',
 
         [ValidateSet('Guerrilla', 'Professional', 'Slate')]
-        [string]$ReportStyle = 'Guerrilla',
+        [string]$ReportStyle = 'Professional',
 
         [switch]$TestMode,